refactor: Moved consts to dedicated file
Consts are also stored in a class now, as opposed to the old dictionary
method. I thought the way to reference dictionary entries was too
clunky, and classes may be more flexible in the end.

Notably, this is part of us trying to get a MVC Model/View/Controller
thing going.