refactor: Changed FileDialogManager to be less stateful
FileDialogManager now doesn't store filepaths as attributes, and its
methods now rather returns them. Yay statelessness!

It now also initializes with an argument for a parent window, and this
attribute is used for its various methods, instead of having to pass it
every method call.

Changing the tests for this was kind of a pain.