fix: Cleared some of the fallout from renaming main_window
It seemed to have triggered some kind of latent bug in the pytests of
stacked layout manager. You need to construct a QApplication before any
QWidgets apparently. We didn't do that before, but the bug triggered
now, so we fixed it. Yeay!

It *was* kind of a pain to find what was wrong, by the way. Next time
use `pytest -s` to disable capturing of error messages and such ^^;