feat: Stupid implementation of proper feature retrieval
We're finally at the stage where we're selecting images for plotting!
Yay! Problem is they take a lot of time to load. I'm sure torch has
multithreading, but we should look into having it at Qt-level as well,
for better usability.

Also note that there's one test which has to go through the entire
loading-images song-and-dance now. So it takes like a minute to finish.
That's bad unit test design! (It's really an oversight)

Updates: ... Issue about this. I can't check cause I'm offline.