Fix pytest hanging when collecting items
I believe the problem was QApplication getting instantiated twice,
though in different programs. Now it's just instantiated in
'__init__.py' instead.

This was tricky because the bug didn't show up in the ghactions, or the
local tests of which. Which was actually a clever hint because those
test through xvfb, which does graphical stuff all in virtual memory,
implying that the problem might have something to do with the windows.
A clever puzzle resolved by a me.