Utilize property decorator for the fcn-resnet-101 class
This was tricky because the @properties decorator utilizes the
'='-operator, which I didn't consider for a while. So I kept calling the
property with the intended values as arguments. Oops!

Besides, now setting the property with the '='-operator also has some
side-effects, which seems like bad practise. Looking into it.