Dialog almost at complete functionality,
As it is, the user can see the entire summary of the model parameters,
and select the layers so that the start-- and end layers will be the
return values of the dialog method.

However, I had the impression that selecting a layer should redraw the
textbox, and this doesn't happen, (why is unbeknownst to me). Also I
don't know what the point of selecting the end layer is. There's no
precedent for it in the rest of the program.

@Linnea, could you take a look at what's here with these points in mind?
Pretty please?

Updates: #25