Fix error where single-category models fail
This was due to `torch.squeeze` removing all dimensions of size 1, such
as in our program, when a tensor has a single category.

The fix was to specify which dimension to remove, such that it won't
erroneously remove a necessary dimension.

This change also improves the associated error message, which will
hopefully be useful for other problems than the one that was just
solved.